project/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
contourpy==1.0.7
cycler==0.11.0
Django==3.2.18
django-admin-interface==0.24.2
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-js-asset==2.0.0
django-seed==0.3.1
django-widget-tweaks==1.4.12
Faker==18.3.1
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.7.0
numpy==1.24.2
packaging==23.0
Pillow==9.0.0
psycopg2==2.9.3
pycodestyle==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
toposort==1.10
tzdata==2021.5
